remove gettext/libintl dependency
gettext is only used for cleos localization support. But, cleos has not had any effort to translate it nor is that effort on any short term roadmap AFAIK. Remove this dependency for now as it's an LGPL dep and causing some grief. When we get back around to localizing cleos we can reevaluate this need; it's possible something like boost::locale would be a better fit.