Fix non-determinism issues with deep-mind logger in tests.
Added an updated configuration `deep_mind_handler` to support
special adjusted behavior for the deep mind logger to be used
specifically in the context of tests.

In particular a new config flag called `zero_elapsed` is added to this
config which if set to true will first zero out the `elapsed` fields
within a copy of the transaction trace before writing it out to the
logs.

The `zero_elapsed` flag is set to true in the deep_mind_tests in order
to guarantee a deterministically generated log to compare against the
committed log. This appears to be a more reliable solution to avoiding
non-determinism in the tests than relying on the auto-generated
`[[:xdigit:]]` masks in the committed logs. It may now be reasonable
to greatly simply that log merge and comparison logic, but that is not
addressed in this commit.