Update appbase submodule commit.
This incorporates the appbase changes that stops the build from breaking
just because this repo does not have any tags.

Note that Leap programs are already dynamically setting the appbase
version strings on startup using their own mechanism, and so the
value of the hardcoded appbase_version_string is irrelevant.