Rework performance_test_basic.py to work as an import module and as a script.
This was necessary as a first step to being able to use this functionality as an import module for an orchestrator to run multiple instances of PerformanceTestBasic with different tps configurations.

Some rework to how args are handled in the report.  When run as an import module, command line args won't be the same or available.  Get args/config info from how the PerformanceTestBasic was configured instead.

Created TestHelperConfig and ClusterConfig objects to help simplify and separate concerns for all the parameters to the test.