For initial binary search, try a short run using ceiling tps first.
In the general case, where --max-tps-to-test has been set well above achievable, this change adds one --test-iteration-duration-sec of time to the overarching test. In the case where it is successful it could save many cycles of time.  It also puts one data point in the results summary that is basically an optimistic execution showing at potential overload what the avg tps looks like.