GH-637 Working to fix nodeos_snapshot_diff_test by sync on trx gen startup.
Refactor - move launchTrxGenerators from Node to Cluster where it fits better.

Add waitForTrxGeneratorsSpinup function which allows tests to wait for initial trx from each generator to appear in a block.  This allows the tests to sync on the trx generators spinning up and knowing that trxs are flowing before proceeding with tests.

Update tests that used launchTrxGenerators for the refactor and to use the waitForTrxGeneratorsSpinup function.

trx_generator now logs first transaction it is preparing to send right before sending it to allow synchronization by tests without influencing its generation rate since it occurs prior only on initil trx and prior to initial trx send.