Remove log_reader_tests.py and associated test log files.
With rework to no longer rely on parsing a majority of the block and transaction data out of the nodeos log in favor of querying it directly from an active node, the log_reader_tests have lost their usefulness.  Can remove nodeos logs as well which were supporting the testing.