Reopen checkOutputFilename in case where file may have been deleted, but attr was already set.
Allows for cases where Cluster cleanup is called initially in tests to remove old test data, which should no longer be a problem, however in doing so it now deletes subprocess_results.log while attr for checkOutputFile has already been set.  This allows reopening/recreating the file and continuing to log to it to capture all commands after the initial cleanup.