Fix issues with processing read-only transactions.
Read only transactions are not logged in trace for cpu usage since they are not billed as such, thus use the elapsed time for the transaction processing to get the cpu usage overall.

Because read only transactions don't make it into blocks in the block log, need to add the transactions from the log file into the data block transaction counter for use in tps stats calculations in performance harness.