Add stack depth check.
For 10 levels of recursion, we used ~1540 bytes of stack.

This translates to 5KB of stack space (32 levels of recursion) for appending 4 billion digests
(the max. possible number of blocks).

Since the default Ubuntu stack for a process is 10MB, and 2MB for a thread, this means that, even
for a thread,  we'd use at most 5/2000, or 0.25% of the stack available for any append().

If the stack is already 99.75% used before `incremental_merkle_tree::append()` is called, we have a
problem elsewhere.